backend/requirements.txt
fastapi==0.110.1
uvicorn
gunicorn
boto3
requests-oauthlib
cryptography
python-dotenv
pymongo>=4.9,<4.10
pydantic>=2.6.4
email-validator
pyjwt
passlib>=1.7.4
tzdata>=2024.2
motor>=3.6.0
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests
pandas
numpy
python-multipart
jq
typer
pillow
python-magic
